% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptLogLik.R
\name{OptLogLik}
\alias{OptLogLik}
\title{OptLogLik (not for call)}
\usage{
OptLogLik(T1, T2, l, knot, p_n, q_n, lower1, upper1, lower2, upper2)
}
\arguments{
\item{T1}{Observed intervals for outcome 1}

\item{T2}{Observed intervals for outcome 2}

\item{l}{Order of I-spline (degree + 1)}

\item{knot}{Knot positions for the I-spines}

\item{p_n}{Number of parameters corresponding to the I-spline basis (for outcome 1)}

\item{q_n}{Number of parameters corresponding to the I-spline basis (for outcome 2)}

\item{lower1}{t1_l}

\item{upper1}{t1_h}

\item{lower2}{t2_l}

\item{upper2}{t2_h}
}
\value{
A list of outcomes including the
\item{rho}{estimated \eqn{\rho} indicating the correlations between two outcomes}
\item{seive.M_ij}{Seive MLE for Mu}
\item{seive.w_i}{Seive MLE for w}
\item{seive.p_j}{Seive MLE for p}
}
\description{
Optimize the loglikelihood of interval censored data
}
