% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDat_case1.R
\docType{data}
\name{SampleDat_case1}
\alias{SampleDat_case1}
\title{Sample toy data to illustrate the package usage}
\format{
A list of 2 items. T1: interval censored data from outcome 1; T2: interval censored data from outcome 2.
}
\description{
A dataset containing 400 observations from two current status data outcomes. The data are generated by the example shown in function \code{SimDat_BiVarIntCensd}.
}
\details{
\itemize{
  \item tu_1, tu_2 left boundary of the observation intervals
  \item tv_1, tv_2 right boundary of the observation intervals
  \item ind_1, ind_2 indicators of censoring types
}
}
